Create prtg-cmuptime-sb8200.pl
This will return the uptime for SB8200 in seconds and PRTG will display in seconds, minutes, hours, days, etc. just like a SNMP uptime sensor.  Lower warning limit is set to 24 hours as I find it's always good to know when something has low uptime.  Helps sometimes as you may not expect the device to be rebooted, so the warning alerts you of this.